Prepare code for easier testing
* Move program logic to main.js

* Move more arguments into parameters
